Fix an int* being used as a void**, which is dubious on LP64 targets.
llvm-svn: 61717

+++ SingleSource/UnitTests/Threads/tls.c
@@ -1,20 +1,21 @@
 #include <stdio.h>
 #include <pthread.h>
+#include <inttypes.h>
 
 void *f(void *a){
-  static __thread int i = 1;
+  static __thread intptr_t i = 1;
   i++;
   return (void *)i;
 }
 
 int main() {
   pthread_t t;
-  int ret;
+  intptr_t ret;
   pthread_create(&t, NULL, f, NULL);
   pthread_join(t, (void **) &ret);
-  printf("Thread 1: %d\n",ret);
+  printf("Thread 1: %" PRIdPTR "\n",ret);
   pthread_create(&t, NULL, f, NULL);
   pthread_join(t, (void **) &ret);
-  printf("Thread 2: %d\n",ret);
+  printf("Thread 2: %" PRIdPTR "\n",ret);
   return 0;
 }
